@@ -1,65 +1,80 @@
 {
   "name": "augustinmauroy.github.io",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
+    "postbuild": "node ./scripts/postbuild.js",
     "start": "next start",
-    "lint-css": "stylelint \"**/*.css\" --allow-empty-input --cache --cache-strategy=content --cache-location=.stylelintcache",
-    "lint-css:fix": "stylelint \"**/*.css\" --allow-empty-input --cache --cache-strategy=content --cache-location=.stylelintcache --fix",
-    "lint-js": "next lint --cache-strategy=content --cache-location=.eslintcache",
-    "lint-js:fix": "npm run lint-js -- --fix",
-    "test": "jest --cache",
-    "test:ci": "jest --ci --coverage --cache",
-    "prettier:write": "prettier --write . --cache --cache-location=.prettiercache",
-    "prettier:check": "prettier --check . --cache --cache-location=.prettiercache",
-    "format": "npm run prettier:write && npm run lint-js:fix && npm run lint-css:fix",
-    "storybook": "storybook dev -p 6006 --quiet --no-open",
-    "storybook:build": "storybook build"
+    "test": "vitest run",
+    "test:watch": "vitest",
+    "test:coverage": "vitest run --coverage",
+    "lint:js": "next lint --cache-strategy=content --cache-location=.eslintjscache",
+    "lint:css": "stylelint \"**/*.css\" --cache --cache-strategy=content --cache-location=.stylelintcache",
+    "prettier": "prettier --write . --cache --cache-strategy=content --cache-location=.prettiercache",
+    "prettier:check": "prettier --check . --cache --cache-strategy=content --cache-location=.prettiercache",
+    "type-check": "tsc --noEmit",
+    "format": "turbo run lint:js lint:css -- --fix && turbo run prettier -- --write",
+    "storybook": "BUILD_ENV=storybook storybook dev -p 6006 --quiet --no-open",
+    "storybook:build": "BUILD_ENV=storybook storybook build",
+    "prepare": "husky"
   },
   "dependencies": {
-    "@radix-ui/react-avatar": "~1.0.4",
-    "bright": "~0.8.5",
+    "@heroicons/react": "~2.1.5",
+    "@radix-ui/react-avatar": "~1.1.0",
+    "@radix-ui/react-dropdown-menu": "~2.1.1",
     "classnames": "~2.5.1",
-    "framer-motion": "^11.2.6",
-    "js-yaml": "~4.1.0",
-    "next": "~14.2.3",
-    "next-intl": "~3.14.1",
+    "feed": "~4.2.2",
+    "geist": "~1.3.1",
+    "next": "~14.2.13",
+    "next-intl": "~3.19.4",
     "next-mdx-remote": "~5.0.0",
-    "next-themes": "~0.3.0",
     "react": "~18.3.1",
     "react-dom": "~18.3.1",
-    "react-icons": "~5.2.1",
-    "turbo": "~1.13.3"
+    "rehype-autolink-headings": "~7.1.0",
+    "rehype-mdx-code-props": "~3.0.1",
+    "rehype-slug": "~6.0.0",
+    "shiki": "~1.18.0",
+    "turbo": "~2.1.2"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "~8.1.3",
-    "@storybook/addon-interactions": "~8.1.3",
-    "@storybook/addon-links": "~8.1.3",
-    "@storybook/addon-themes": "~8.1.3",
-    "@storybook/nextjs": "~8.1.3",
-    "@storybook/react": "~8.1.1",
-    "@testing-library/jest-dom": "~6.4.5",
-    "@testing-library/react": "~15.0.7",
-    "@types/jest": "~29.5.12",
-    "@types/node": "~20",
-    "@types/react": "~18.3.3",
+    "@storybook/addon-a11y": "~8.3.2",
+    "@storybook/addon-essentials": "~8.3.2",
+    "@storybook/addon-interactions": "~8.3.2",
+    "@storybook/addon-links": "~8.3.2",
+    "@storybook/addon-themes": "~8.3.2",
+    "@storybook/nextjs": "~8.3.2",
+    "@storybook/react": "~8.3.2",
+    "@testing-library/dom": "~10.4.0",
+    "@testing-library/react": "~16.0.1",
+    "@types/node": "~20.13.0",
+    "@types/react": "~18.3.8",
     "@types/react-dom": "~18.3.0",
-    "autoprefixer": "~10.4.19",
-    "eslint": "~8",
-    "eslint-config-next": "~14.2.3",
+    "@vitejs/plugin-react": "~4.3.1",
+    "autoprefixer": "~10.4.20",
+    "eslint": "~8.57.1",
+    "eslint-config-next": "~14.2.13",
+    "eslint-config-prettier": "~9.1.0",
     "eslint-plugin-storybook": "~0.8.0",
-    "jest": "~29.7.0",
-    "jest-environment-jsdom": "~29.7.0",
-    "postcss": "~8",
-    "prettier": "~3.2.5",
-    "prettier-plugin-tailwindcss": "~0.5.14",
-    "storybook": "~8.1.3",
-    "stylelint-config-standard": "^36.0.0",
-    "stylelint-order": "^6.0.4",
-    "stylelint-selector-bem-pattern": "^4.0.0",
-    "tailwindcss": "~3.4.3",
-    "typescript": "~5"
-  }
+    "husky": "~9.1.6",
+    "jsdom": "~25.0.1",
+    "lint-staged": "~15.2.10",
+    "postcss": "~8.4.47",
+    "prettier": "~3.3.3",
+    "prettier-plugin-tailwindcss": "~0.6.6",
+    "storybook": "~8.3.2",
+    "stylelint-config-standard": "~36.0.1",
+    "stylelint-order": "~6.0.4",
+    "stylelint-selector-bem-pattern": "~4.0.1",
+    "tailwindcss": "~3.4.11",
+    "typescript": "~5.6.2",
+    "vitest": "~2.1.1"
+  },
+  "engines": {
+    "node": "20.x",
+    "npm": "10.x"
+  },
+  "packageManager": "npm@10.0.0"
 }
