@@ -1,3 +1,4 @@
+/** @type {import('postcss-load-config').Config} */
 const config = {
   plugins: {
     'tailwindcss/nesting': {},
@@ -6,4 +7,4 @@ const config = {
   },
 };
 
-module.exports = config;
+export default config;
