@@ -8,34 +8,38 @@ const config: StorybookConfig = {
     '@storybook/addon-essentials',
     '@storybook/addon-themes',
     '@storybook/addon-interactions',
+    '@storybook/addon-a11y',
   ],
   logLevel: 'error',
   staticDirs: ['../public'],
   typescript: { reactDocgen: false, check: false },
-  core: { disableTelemetry: true, disableWhatsNewNotifications: true },
+  docs: {
+    autodocs: false,
+    docsMode: false,
+  },
+  core: {
+    disableTelemetry: true,
+    disableWhatsNewNotifications: true,
+  },
   framework: {
     name: '@storybook/nextjs',
     options: { builder: { useSWC: true } },
   },
   previewBody:
-    // This `<style>` is necessary to simulate what `next-themes` (ThemeProvider) does on real applications
-    // `next-theme` automatically injects the color-scheme based on the system preference or the current applied theme
-    // on Storybook we don't use `next-theme` as we want to simulate themes
     '<style>:root { color-scheme: light; } html[data-theme="dark"] { color-scheme: dark; }</style>' +
-    // This adds the base styling for dark/light themes within Storybook. This is a Storybook-only style
     // Warning: this should be same as the one in `src/styles/globals.css`
-    '<body class="bg-gray-50 text-gray-950 dark:bg-gray-900 transition-colors dark:text-gray-50"></body>',
+    '<body class="bg-yellow-50 text-gray-950 dark:bg-gray-950 dark:text-gray-50"></body>',
   webpack: async config => ({
     ...config,
     // Performance Hints do not make sense on Storybook as it is bloated by design
     performance: { hints: false },
     // Removes Pesky Critical Dependency Warnings due to `next/font`
     ignoreWarnings: [e => e.message.includes('Critical dep')],
-    // allows to use `@/` as a shortcut to the `src/` folder
+    // allows to use `~/` as a shortcut to the `src/` folder
     resolve: {
       ...config.resolve,
       // @ts-ignore
-      alias: { ...config.resolve.alias, '@': resolve(__dirname, '../src') },
+      alias: { ...config.resolve.alias, '~': resolve(__dirname, '../') },
     },
   }),
 };
